Add a section for the development model
Add a section in chapter 3 that discusses the development model that the
Eventify system will use.

Also renamed the chapter to "Development Model" instead of the previous
"Proposed Model" to better reflect the content of the chapter.

Renamed the section on UML diagrams from "Development Diagrams" to
"System Design" to avoid redundant language.