Fix ticket assignee details not updating on attendee page
A bug was fixed where a ticket assignee's details were not updated on
the attendee page after the user updated their details. This was as a
result of a redundant write to the database.

The bug was fixed by on the API and the necessary changes were made on
the frontend to ensure the attendee page is updated with the correct
details.